+++ docs-src/components/push-button/PushButton.svelte
@@ -4,24 +4,27 @@
 
 <h4>Default</h4>
 <PushButton>Hello</PushButton>
+<PushButton info>Info</PushButton>
 <PushButton success>Warning</PushButton>
 <PushButton warning>Warning</PushButton>
 <PushButton danger>Danger</PushButton>
 
 <h4>Pressed</h4>
 <PushButton pressed>Hello</PushButton>
+<PushButton pressed info>Info</PushButton>
 <PushButton pressed success>Warning</PushButton>
 <PushButton pressed warning>Warning</PushButton>
 <PushButton pressed danger>Danger</PushButton>
 
 <h4>Disabled</h4>
 <PushButton disabled>Hello</PushButton>
+<PushButton disabled info>Info</PushButton>
 <PushButton disabled success>Success</PushButton>
 <PushButton disabled warning>Warning</PushButton>
 <PushButton disabled danger>Danger</PushButton>
 
 <h4>With icon</h4>
-<PushButton icon="info">Info</PushButton>
+<PushButton icon="info" info>Info</PushButton>
 <PushButton icon="check" success>Success</PushButton>
 <PushButton icon="alert" warning>Warning</PushButton>
 <PushButton icon="trash" danger>Delete</PushButton>
@@ -31,7 +34,7 @@
 <h3>Icon only buttons</h3>
 
 <h4>Default</h4>
-<PushButton icon="info"></PushButton>
+<PushButton icon="info" info></PushButton>
 <PushButton icon="check" success></PushButton>
 <PushButton icon="alert" warning></PushButton>
 <PushButton icon="trash" danger></PushButton>
@@ -41,7 +44,7 @@
 <h3>Icon only, and round</h3>
 
 <h4>Default</h4>
-<PushButton round icon="info"></PushButton>
+<PushButton round icon="info" info></PushButton>
 <PushButton round icon="check" success></PushButton>
 <PushButton round icon="alert" warning></PushButton>
 <PushButton round icon="trash" danger></PushButton>